requirements.txt
tensorflow
numpy
Pillow
gunicorn
flask
